refactor neb diffusion
