+++ src/simmate/calculators/vasp/workflows/diffusion/README.md
