+++ src/simmate/calculators/vasp/workflows/diffusion/_to_do/from_startend_sites.py
@@ -1,0 +1,1 @@
+# -*- coding: utf-8 -*-
