+++ src/simmate/calculators/vasp/workflows/diffusion/test/all_paths/bulk_relaxation/vasprun.xml
