+++ src/simmate/calculators/vasp/workflows/diffusion/test/all_paths/bulk_static_energy/vasprun.xml
