+++ src/simmate/calculators/vasp/workflows/diffusion/test/all_paths/migration_hop_01/endpoint_relaxation_start/OUTCAR
