+++ src/simmate/calculators/vasp/workflows/diffusion/test/test_all_paths.py
