--- src/simmate/calculators/vasp/workflows/nudged_elastic_band/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .all_paths import workflow as all_paths_workflow
-from .single_path import workflow as single_path_workflow
-from .from_endpoints import workflow as from_endpoints_workflow
-from .from_images import workflow as from_images_workflow
