+++ src/simmate/workflows/diffusion.py
@@ -4,9 +4,9 @@
 Workflows for analyzing ionic diffusion with a material
 """
 
-from simmate.calculators.vasp.workflows.nudged_elastic_band import (
-    all_paths_workflow,
-    single_path_workflow,
-    from_endpoints_workflow,
-    from_images_workflow,
+from simmate.calculators.vasp.workflows.diffusion.all import (
+    Diffusion__Vasp__NebAllPaths,
+    Diffusion__Vasp__NebSinglePath,
+    Diffusion__Vasp__NebFromEndpoints,
+    Diffusion__Vasp__NebFromImages,
 )
